+++ starships/src/main/scala/entity/value/motion/Motion.scala
@@ -3,7 +3,7 @@ package entity.value.motion
 import java.lang.Math.sin
 import java.lang.Math.cos
 
-case class Motion(position: Coordinate, degree: Int, speed: Double):
+case class Motion(position: Coordinate, degree: Double, speed: Double):
   def newMotion: Motion =
     val newCoordinate = Coordinate(position.x + speed * sin(Math.toRadians(degree))
       , position.x - speed * cos(Math.toRadians(degree)))
